requirements.txt
requests
kafka-python
configparser
pyspark
